CRUD 03122020

+++ Pertemuan 10 PHP (CRUD)/hapus_biodata.php
@@ -1,7 +1,13 @@
 <?php 
+session_start();
+if (!isset($_SESSION["password"])) {
+  header("Location:login.php");
+  exit;
+}
+
 include 'koneksi.php';
 $id = $_GET['id'];
 
 $sql = $koneksi->query("DELETE FROM biodata WHERE id='$id'");
 
-header('Location:index.php');
+header('Location:biodata.php');
